fixing a bug that rewrites the output file

+++ src/pkd/extract_genes.py
@@ -15,8 +15,12 @@ def extract_variants(input):
                     if not os.path.exists(f"/data/project/worthey_lab/projects/experimental_pipelines/tarun/ditto/data/processed/dbnsfp_genes/{gene}"):
                         os.makedirs(f"/data/project/worthey_lab/projects/experimental_pipelines/tarun/ditto/data/processed/dbnsfp_genes/{gene}")
 
-                    with gzip.open(f"/data/project/worthey_lab/projects/experimental_pipelines/tarun/ditto/data/processed/dbnsfp_genes/{gene}/dbNSFP_{gene}_variants.tsv.gz", "wt") as out:
-                        out.write(line + "\n")
+                    if os.path.isfile(f"/data/project/worthey_lab/projects/experimental_pipelines/tarun/ditto/data/processed/dbnsfp_genes/{gene}/dbNSFP_{gene}_variants.tsv.gz"):
+                        with gzip.open(f"/data/project/worthey_lab/projects/experimental_pipelines/tarun/ditto/data/processed/dbnsfp_genes/{gene}/dbNSFP_{gene}_variants.tsv.gz", "at") as out:
+                            out.write(line + "\n")
+                    else:
+                        with gzip.open(f"/data/project/worthey_lab/projects/experimental_pipelines/tarun/ditto/data/processed/dbnsfp_genes/{gene}/dbNSFP_{gene}_variants.tsv.gz", "wt") as out:
+                            out.write(line + "\n")
 
     return None
